+++ lib/src/dart/enums.dart
@@ -1,6 +1,11 @@
 /// 定位模式
 enum LocationAccuracy {
+  /// 高
   High,
+
+  /// 低
   Low,
-//  DeviceSensor,
+
+  /// 设备传感器 仅android
+  DeviceSensor,
 }